Update pointer diff types

+++ src/mobius/Stream.cpp
@@ -486,8 +486,8 @@ PUBLIC int Stream::getPitchTweak(int tweak)
 
 PUBLIC long Stream::deltaFrames(float* start, float* end)
 {
-	ptrdiff_t bytes = end - start;
-	long samples = (long) bytes / sizeof(float);
+	long bytes = (long)((ptrdiff_t)end - (ptrdiff_t)start);
+	long samples = bytes / sizeof(float);
 	long frames = samples / channels;
 	return frames;
 }
